+++ src/oihana/memcached/enums/MemcachedDefinition.php
@@ -16,9 +16,9 @@ class MemcachedDefinition
     use ConstantsTrait ;
 
     const string APP_PATH        = 'appPath'         ;
-    const string CACHE_MEMCACHED = 'appPath'         ;
+    const string CACHE_MEMCACHED = 'cache:memcached' ;
     const string CACHE_MEMORY    = 'cache:memory'    ;
-    const string COMMAND         = 'cache:memcached' ;
+    const string COMMAND         = 'command'         ;
     const string COMMANDS        = 'commands'        ;
     const string CONFIG          = 'config'          ;
     const string CONFIG_PATH     = 'configPath'      ;